+++ objectified-ui/public/WHATS_NEW.md
@@ -56,6 +56,7 @@ We continue to improve the platform based on your feedback with improvements and
   - Added indicator to paths that are invalid (or not completely defined)
   - Added display of sample URL for each path that is created in the canvas
   - Added content type definitions for request body schemas
+  - Added display of description to the request body schema definitions
   - Dragging a property to a variable chip in the header will apply that property's primitive settings to the bound variable
   - Dragging a class to a response schema allows the class to be copied or referenced
 - OpenAPI Support: